Refactorizacion carpetas y nombres
