+++ src/main/java/com/eoi/springboot/ControladorUsuarios.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.eoi.springboot;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
