+++ src/main/java/com/eoi/springboot/MySpringBootApplication.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.eoi.springboot;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,10 +6,10 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication // Haz tu magia, SPRING
 
-public class DemoApplication {
+public class MySpringBootApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(DemoApplication.class, args);
+		SpringApplication.run(MySpringBootApplication.class, args);
 	}
