+++ src/test/java/com/eoi/springboot/MySpringBootApplicationTests.java
@@ -1,10 +1,10 @@
-package com.example.demo;
+package com.eoi.springboot;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class DemoApplicationTests {
+class MySpringBootApplicationTests {
 
 	@Test
 	void contextLoads() {
